requirements.txt
streamlit
torch
transformers
Pillow
numpy
torchvision
